# docs/swagger.yaml
openapi: 3.0.0
info:
  title: "UNAHUR Anti-social Net API"
  description: "API de la red social Anti-social Net para la Universidad Nacional de Hurlingham (UNAHUR)."
  version: "1.1.0"
servers:
  - url: "http://localhost:4002"
    description: "Servidor local de desarrollo"
paths:
...
  /users/{id}/posts:
    get:
      summary: "Obtener posts de un usuario"
      description: "Retorna todos los posts publicados por el usuario especificado."
      tags:
        - Users
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          description: "ID del usuario"
      responses:
        '200':
          description: "Posts devueltos exitosamente"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '400':
          description: "ID inválido"
        '404':
          description: "Usuario no encontrado"
...
  /posts:
    get:
      summary: "Obtener todos los posts"
      description: "Retorna todos los posts ordenados por fecha de creación descendente, incluyendo tags, imágenes y comentarios no vencidos."
      tags:
        - Posts
      responses:
        '200':
          description: "Lista de posts devuelta exitosamente"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
    post:
      summary: "Crear un nuevo post"
      description: "Crea una nueva publicación para un usuario."
      tags:
        - Posts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: "Post creado exitosamente"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: "Error de validación o ID de usuario inexistente"
...
  /posts/{id}:
    get:
      summary: "Obtener un post por ID"
      description: "Retorna los detalles de un post específico."
      tags:
        - Posts
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          description: "ID del Post"
      responses:
        '200':
          description: "Post obtenido exitosamente"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: "Post no encontrado"

    put:
      summary: "Actualizar un post por ID"
      description: "Modifica el contenido de un post existente."
      tags:
        - Posts
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          description: "ID del Post"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '200':
          description: "Post actualizado exitosamente"
        '404':
          description: "Post no encontrado"

    delete:
      summary: "Eliminar un post por ID"
      description: "Elimina permanentemente un post."
      tags:
        - Posts
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
          description: "ID del Post"
      responses:
        '200':
          description: "Post eliminado con éxito"
        '404':
          description: "Post no encontrado"
...
  /posts/{id}/image:
    post:
      summary: "Agregar imagen a un post"
      description: "Asocia una nueva URL de imagen al post especificado."
      tags:
        - Posts
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
              properties:
                url:
                  type: string
                  format: uri
                  example: "https://ejemplo.com/imagen.jpg"
      responses:
        '201':
          description: "Imagen asociada correctamente"
        '404':
          description: "Post no encontrado"
...
  /comments/post/{postId}:
    get:
      summary: "Obtener todos los comentarios activos de un post por su ID"
      tags:
        - Comments
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '200':
          description: "Comentarios no vencidos del post"

  # -------------------- TAGS --------------------
...
  /postImages:
    get:
      summary: "Obtener todas las imágenes de posts"
      tags:
        - Post Images
      responses:
        '200':
          description: "Lista de imágenes obtenida"
    post:
      summary: "Crear registro de imagen (Carga física mediante Multer)"
      description: "Permite subir una imagen utilizando multipart/form-data. El archivo debe enviarse en la propiedad 'url' y el ID del post en 'post'."
      tags:
        - Post Images
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - url
                - post
              properties:
                url:
                  type: string
                  format: binary
                  description: "Archivo de imagen físico"
                post:
                  type: string
                  pattern: '^[0-9a-fA-F]{24}$'
                  description: "ID del post asociado"
      responses:
        '201':
          description: "Imagen subida e ingresada con éxito"
  /post_Images:
    get:
      summary: "Obtener todas las imágenes de posts (Alias)"
      tags:
        - Post Images
      responses:
        '200':
          description: "Lista de imágenes"
    post:
      summary: "Crear registro de imagen (Alias)"
      tags:
        - Post Images
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - url
                - post
              properties:
                url:
                  type: string
                  format: binary
                post:
                  type: string
                  pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '201':
          description: "Imagen creada"
  /postImages/{id}:
    get:
      summary: "Obtener una imagen por ID"
      tags:
        - Post Images
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '200':
          description: "Imagen devuelta"
    put:
      summary: "Actualizar URL de imagen"
      tags:
        - Post Images
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
              properties:
                url:
                  type: string
                  format: uri
      responses:
        '200':
          description: "Registro actualizado"
    delete:
      summary: "Eliminar una imagen por ID"
      description: "Elimina el registro de la base de datos y borra el archivo del sistema de archivos local."
      tags:
        - Post Images
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '200':
          description: "Imagen eliminada exitosamente"
  /post_Images/{id}:
    get:
      summary: "Obtener una imagen por ID (Alias)"
      tags:
        - Post Images
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '200':
          description: "Imagen devuelta"
    delete:
      summary: "Eliminar una imagen por ID (Alias)"
      tags:
        - Post Images
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
      responses:
        '200':
          description: "Imagen eliminada"
components:
  schemas:
...
    Post:
      type: object
      required:
        - id
        - description
        - user
      properties:
        id:
          type: string
          example: "60c72b2f9b1d8b3c6c8e4f11"
        description:
          type: string
          example: "Este es mi primer post en la red antisocial"
        user:
          type: object
          properties:
            id:
              type: string
            nickName:
              type: string
        tags:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
        images:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
    CreatePostRequest:
      type: object
      required:
        - description
        - user
      properties:
        description:
          type: string
          minLength: 1
          maxLength: 500
          example: "Este es el contenido del post"
        user:
          type: string
          pattern: '^[0-9a-fA-F]{24}$'
          example: "60c72b2f9b1d8b3c6c8e4f10"
        tags:
          type: array
          items:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
    Comment:
      type: object
      required:
        - id
        - text
        - post
        - user
      properties:
        id:
          type: string
          example: "60c72b2f9b1d8b3c6c8e4f12"
        text:
          type: string
          example: "Este es un comentario de prueba"
        post:
          type: string
        user:
          type: object
          properties:
            id:
              type: string
            nickName:
              type: string
        createdAt:
          type: string
...